requirements.txt
fastapi
uvicorn
alembic
sqlalchemy
psycopg[binary]
python-jose[cryptography]
bcrypt
pydantic[email]